Enhanced UserPreferenceManager.js to support multiple NG-CHMs
When opened, the preferences manager will be for the current heat map.

If the preferences manager is open when a new heatMap is selected, it
will stay open for the old map and any updates will still apply to that
map.

If the user re-opens the preferences manager when it is already open, the
open preferences manager will close and a new preferences manager for the
current heat map will open.