Removed dependency of SearchState.js on MMGR
The core of this change are the updates to SearchState.js.  MMGR was
removed as a dependency of the module, and the functions that previously
used MMGR.getHeatMap() to obtain a heatMap now accept it as a parameter.

All the other updates are adding a heatMap to the parameters of calls to
these functions.

This is in preparation for changing from a global search state to a search
state per heatMap.